feat: show real auto-detected icon in group dialog preview
Centralize the pattern-based auto-detect logic in GroupIconResolver and
reuse it from ProjectFileGroupNode, the settings table, and the Edit
File Group dialog. The dialog preview now reflects the icon that will
actually render in the tree (e.g. Kotlin icon for *.kt) instead of a
hardcoded folder, and refreshes as the user edits patterns.
